Refactor codebase.
Improvements made
 - Create game in server.js now
 - Remove modal overlay look
 - Hotswap roles now
 - SVG text for cards
 - Remove client from list on disconnect
