--- proposal.md
@@ -1,22 +1,0 @@
-# cs4241-FinalProject
-
-## Team Members
-- Daniel Duff
-- Timothy Winters
-- Andrew Levy
-
-_Online **Codenames** Multiplayer Game_
-For our Final Project, we plan on implementing an online version of the card game _codenames_ which will allow users to play together online and create their own cards and play styles.
-
-## This project will feature:
-- HTML static pages and validations for form input to prevent duplicate cards
-- A login feature, which allows users to log in and edit/produce their own personal card decks.
-- A Node.js SQLite database which hold all of the login and card deck info for all users.
-- Dynamic Javascript that will implement selectors and buttons for selecting and making cards in game as well as communicating with teammates (if necessary) 
-- Javascript to interact (add/modify/remove) with the database. 
-- Card presets, which will allow users to play the game without needing to create their own deck (if they so choose)
-
-## Goals for this Project:
-- Create a fun interactive game that people can play online.
-- Utilize web sockets to allow for realtime communication between players, browser, and the server.
-- Allow players customization options for the game that paying with cards in real life does not offer.
